requirements.txt
flask==3.1.3
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==12.1.1
gunicorn
requests
